handling direct systemcalls and appyling x86 archtecture with cpu feature selection for optimization

+++ res_man/usage/parser/cpu_parser.h
@@ -1,2 +1,2 @@
-__attribute__((hot)) float numOfCores(char* __restrict__ data, char* __restrict__ uptimeData);
-__attribute__((hot)) float formatCpu(int a, int b);
+__attribute__((hot)) double numOfCores(char* __restrict__ data, char* __restrict__ uptimeData);
+__attribute__((hot)) double formatCpu(int a, int b);
